# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"WeatherComfort Analytics API","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Get Health","operationId":"get_health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/v1/weather/comfort":{"get":{"summary":"Get Comfort Weather","operationId":"get_comfort_weather_api_v1_weather_comfort_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComfortWeatherResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/weather/cities/{city_id}/temperature-trend":{"get":{"summary":"Get Temperature Trend","operationId":"get_temperature_trend_api_v1_weather_cities__city_id__temperature_trend_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"city_id","in":"path","required":true,"schema":{"type":"integer","title":"City Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TemperatureTrendResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/cache/status":{"get":{"summary":"Get Cache Status","operationId":"get_cache_status_api_v1_cache_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CacheStatusResponse"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"CacheStatusResponse":{"properties":{"processed_cache":{"$ref":"#/components/schemas/ProcessedCacheStatus"},"raw_cache_summary":{"$ref":"#/components/schemas/RawCacheSummary"}},"type":"object","required":["processed_cache","raw_cache_summary"],"title":"CacheStatusResponse"},"CacheSummary":{"properties":{"processed":{"type":"string","title":"Processed","description":"Processed cache status (e.g., HIT, MISS)"},"raw_hits":{"type":"integer","title":"Raw Hits","description":"Number of raw cache hits"},"raw_misses":{"type":"integer","title":"Raw Misses","description":"Number of raw cache misses"}},"type":"object","required":["processed","raw_hits","raw_misses"],"title":"CacheSummary"},"CityResult":{"properties":{"city_id":{"type":"integer","title":"City Id","description":"Predefined city numeric ID"},"city_name":{"type":"string","title":"City Name","description":"Name of the city"},"country":{"type":"string","title":"Country","description":"Two-letter country code"},"rank":{"type":"integer","title":"Rank","description":"City rank based on overall comfort score (starting at 1)"},"comfort_score":{"type":"number","title":"Comfort Score","description":"Custom Comfort Index score (0.00 - 100.00)"},"weather":{"$ref":"#/components/schemas/WeatherDetails"},"score_breakdown":{"$ref":"#/components/schemas/ScoreBreakdown"}},"type":"object","required":["city_id","city_name","country","rank","comfort_score","weather","score_breakdown"],"title":"CityResult"},"ComfortWeatherResponse":{"properties":{"generated_at":{"type":"string","title":"Generated At","description":"ISO 8601 UTC timestamp when response was generated"},"formula_version":{"type":"string","title":"Formula Version","description":"Comfort score formula version (e.g. v1, v2)"},"city_count":{"type":"integer","title":"City Count","description":"Number of cities successfully processed"},"failed_city_count":{"type":"integer","title":"Failed City Count","description":"Number of cities that failed during fetching"},"cache":{"$ref":"#/components/schemas/CacheSummary"},"cities":{"items":{"$ref":"#/components/schemas/CityResult"},"type":"array","title":"Cities","description":"List of ranked city comfort scores"}},"type":"object","required":["generated_at","formula_version","city_count","failed_city_count","cache","cities"],"title":"ComfortWeatherResponse"},"ForecastPoint":{"properties":{"timestamp_utc":{"type":"string","title":"Timestamp Utc","description":"ISO 8601 UTC timestamp of the forecast point"},"temperature_c":{"type":"number","title":"Temperature C","description":"Temperature in Celsius"}},"type":"object","required":["timestamp_utc","temperature_c"],"title":"ForecastPoint"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"service":{"type":"string","title":"Service"}},"type":"object","required":["status","service"],"title":"HealthResponse"},"ProcessedCacheStatus":{"properties":{"key":{"type":"string","title":"Key","description":"Redis key used for processed cache storage"},"status":{"type":"string","title":"Status","description":"Processed cache status (e.g., HIT, MISS)"},"ttl_seconds":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Ttl Seconds","description":"Time to live in seconds"}},"type":"object","required":["key","status"],"title":"ProcessedCacheStatus"},"RawCacheSummary":{"properties":{"hits":{"type":"integer","title":"Hits","description":"Total raw cache hits across cities"},"misses":{"type":"integer","title":"Misses","description":"Total raw cache misses across cities"}},"type":"object","required":["hits","misses"],"title":"RawCacheSummary"},"ScoreBreakdown":{"properties":{"temperature":{"type":"number","title":"Temperature","description":"Comfort score component for temperature (0-100)"},"humidity":{"type":"number","title":"Humidity","description":"Comfort score component for humidity (0-100)"},"wind":{"type":"number","title":"Wind","description":"Comfort score component for wind speed (0-100)"},"cloudiness":{"type":"number","title":"Cloudiness","description":"Comfort score component for cloudiness (0-100)"}},"type":"object","required":["temperature","humidity","wind","cloudiness"],"title":"ScoreBreakdown"},"TemperatureTrendResponse":{"properties":{"city_id":{"type":"integer","title":"City Id","description":"Predefined city numeric ID"},"city_name":{"type":"string","title":"City Name","description":"Name of the city"},"source_interval_hours":{"type":"integer","title":"Source Interval Hours","description":"Forecast data resolution in hours","default":3},"forecast_points":{"items":{"$ref":"#/components/schemas/ForecastPoint"},"type":"array","title":"Forecast Points","description":"List of forecast temperature data points"},"generated_at":{"type":"string","title":"Generated At","description":"ISO 8601 UTC timestamp when response was generated"},"cache":{"$ref":"#/components/schemas/TrendCacheInfo","description":"Cache information"}},"type":"object","required":["city_id","city_name","forecast_points","generated_at","cache"],"title":"TemperatureTrendResponse"},"TrendCacheInfo":{"properties":{"status":{"type":"string","title":"Status","description":"Cache status (HIT, MISS)"},"ttl_seconds":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Ttl Seconds","description":"Time to live in seconds"}},"type":"object","required":["status"],"title":"TrendCacheInfo"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WeatherDetails":{"properties":{"description":{"type":"string","title":"Description","description":"Weather description text (e.g. overcast clouds)"},"temperature_c":{"type":"number","title":"Temperature C","description":"Temperature in Celsius"},"humidity":{"type":"integer","title":"Humidity","description":"Humidity percentage (0-100)"},"pressure_hpa":{"type":"integer","title":"Pressure Hpa","description":"Atmospheric pressure in hPa"},"wind_speed_mps":{"type":"number","title":"Wind Speed Mps","description":"Wind speed in meters per second"},"cloudiness_percent":{"type":"integer","title":"Cloudiness Percent","description":"Cloudiness percentage (0-100)"},"visibility_km":{"type":"number","title":"Visibility Km","description":"Visibility distance in kilometers"}},"type":"object","required":["description","temperature_c","humidity","pressure_hpa","wind_speed_mps","cloudiness_percent","visibility_km"],"title":"WeatherDetails"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}